Only merge HEAD into deploy if necessary
If the destination branch is already up to date, there is no need to
sync the branch. This is the case when the branch is newly created or
when there have been no new commits since the previous merge.

`git merge-base` is silently used behind the scenes. If it fails it
is the same as no common merge history found, which will trigger a
merge which will either succeed, or fail with a reasonable error
message.