fix callback exception catching
If the next function in the vasync pipeline is what throws the
exception then it is caught by the previous function in the
pipeline and it re-calls its callback, which in turn triggers a
vasync exception that masks the original error.

This is particularly annoying with assert based tests since the test
failure is turned into a vasync error.