remove support for --bundle flag
Use strong-pack instead of 'npm pack', which basically does the same
thing but without any filtering or requirement to use
bundledDependencies in order to include dependencies in the resulting
tarball.

The option is now a no-op since it is essentially always on, but is
left in the flags list so that we don't break any scripts that may be
relying on it.